requirements.txt
arrow==1.2.3
asgiref==3.5.2
blessed==1.19.1
coverage==6.5.0
Django==3.2.16
django-filter==22.1
django-haystack==3.1.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.12.4
drf-haystack==1.8.11
Markdown==3.4.1
python-dateutil==2.8.2
pytz==2022.4
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
wcwidth==0.2.5
whitenoise==6.2.0
Whoosh==2.7.4
